write_record method
